rename supabase/posts.ts -> ads.ts
